decoration code refactor
